Fix test build and make output compatible with current dist.
TODO: revisit exports and dist. Are there really any consumers besides Ember?
