+++ package.json
@@ -32,5 +32,8 @@
     "loader.js": "^3.3.0",
     "qunitjs": "^1.19.0",
     "typescript": "^2.1.6"
+  },
+  "dependencies": {
+    "@types/qunit": "1"
   }
 }
