+++ tests/defer-test.ts
@@ -1,6 +1,11 @@
 import Backburner from 'backburner';
+var originalDateValueOf = Date.prototype.valueOf;
 
-module('deferOnce');
+QUnit.module('defer', {
+  teardown: function(){
+    Date.prototype.valueOf = originalDateValueOf;
+  }
+});
 
 test('when passed a function', function() {
   expect(1);
@@ -9,7 +14,7 @@ test('when passed a function', function() {
   var functionWasCalled = false;
 
   bb.run(function() {
-    bb.deferOnce('one', function() {
+    bb.defer('one', function() {
       functionWasCalled = true;
     });
   });
@@ -24,7 +29,7 @@ test('when passed a target and method', function() {
   var functionWasCalled = false;
 
   bb.run(function() {
-    bb.deferOnce('one', {zomg: 'hi'}, function() {
+    bb.defer('one', {zomg: 'hi'}, function() {
       equal(this.zomg, 'hi', 'the target was properly set');
       functionWasCalled = true;
     });
@@ -47,7 +52,7 @@ test('when passed a target and method name', function() {
   };
 
   bb.run(function() {
-    bb.deferOnce('one', targetObject, 'checkFunction');
+    bb.defer('one', targetObject, 'checkFunction');
   });
 
   ok(functionWasCalled, 'function was called');
@@ -65,7 +70,7 @@ test('throws when passed a null method', function() {
   });
 
   bb.run(function() {
-    bb.deferOnce('deferErrors', {zomg: 'hi'}, null);
+    bb.defer('deferErrors', {zomg: 'hi'}, null);
   });
 });
 
@@ -81,7 +86,7 @@ test('throws when passed an undefined method', function() {
   });
 
   bb.run(function() {
-    bb.deferOnce('deferErrors', {zomg: 'hi'}, undefined);
+    bb.defer('deferErrors', {zomg: 'hi'}, undefined);
   });
 });
 
@@ -97,7 +102,7 @@ test('throws when passed an method name that does not exists on the target', fun
   });
 
   bb.run(function() {
-    bb.deferOnce('deferErrors', {zomg: 'hi'}, 'checkFunction');
+    bb.defer('deferErrors', {zomg: 'hi'}, 'checkFunction');
   });
 });
 
@@ -108,7 +113,7 @@ test('when passed a target, method, and arguments', function() {
   var functionWasCalled = false;
 
   bb.run(function() {
-    bb.deferOnce('one', {zomg: 'hi'}, function(a, b, c) {
+    bb.defer('one', {zomg: 'hi'}, function(a, b, c) {
       equal(this.zomg, 'hi', 'the target was properly set');
       equal(a, 1, 'the first arguments was passed in');
       equal(b, 2, 'the second arguments was passed in');
@@ -121,105 +126,84 @@ test('when passed a target, method, and arguments', function() {
 });
 
 test('when passed same function twice', function() {
-  expect(2);
-
-  var bb = new Backburner(['one']);
-  var i = 0;
-  var functionWasCalled=false;
-  var deferMethod = function(){
-    i++;
-    equal(i, 1, 'Function should be called only once');
-    functionWasCalled = true;
-  };
-
-  bb.run(function() {
-    bb.deferOnce('one', deferMethod);
-    bb.deferOnce('one', deferMethod);
-  });
-
-  ok(functionWasCalled, 'function was called only once');
-});
-
-test('when passed same function twice with same target', function() {
-  expect(3);
-
-  var bb = new Backburner(['one']);
-  var i=0;
-  var functionWasCalled=false;
-  var deferMethod = function(){
-    i++;
-    equal(i, 1, 'Function should be called only once');
-    equal(this['first'], 1, 'the target property was set');
-    functionWasCalled = true;
-  };
-  var argObj = {'first': 1};
-
-  bb.run(function() {
-    bb.deferOnce('one', argObj, deferMethod);
-    bb.deferOnce('one', argObj, deferMethod);
-  });
-
-  ok(functionWasCalled, 'function was called only once');
-});
-
-test('when passed same function twice with different targets', function() {
-  expect(3);
+  expect(1);
 
   var bb = new Backburner(['one']);
   var i = 0;
   var deferMethod = function(){
     i++;
-    equal(this['first'], 1, 'the target property was set');
   };
 
   bb.run(function() {
-    bb.deferOnce('one', {'first': 1}, deferMethod);
-    bb.deferOnce('one', {'first': 1}, deferMethod);
+    bb.defer('one', deferMethod);
+    bb.defer('one', deferMethod);
   });
 
   equal(i, 2, 'function was called twice');
 });
 
-test('when passed same function twice with same arguments and same target', function() {
-  expect(4);
+
+test('when passed same function twice with arguments', function() {
+  expect(2);
 
   var bb = new Backburner(['one']);
-  var i = 0;
-  var deferMethod = function(a, b){
-    i++;
-    equal(a, 1, 'First argument is set only one time');
-    equal(b, 2, 'Second argument remains same');
-    equal(this['first'], 1, 'the target property was set');
-  };
-  var argObj = {'first': 1};
+  var i=0;
+  var deferMethod = function(){
+        equal(this['first'], 1, 'the target property was set');
+      };
+  var argObj = {'first' : 1};
 
   bb.run(function() {
-    bb.deferOnce('one', argObj, deferMethod, 1, 2);
-    bb.deferOnce('one', argObj, deferMethod, 1, 2);
+    bb.defer('one', argObj, deferMethod);
+    bb.defer('one', argObj, deferMethod);
   });
 
-  equal(i, 1, 'function was called once');
 });
 
-test('when passed same function twice with same target and different arguments', function() {
-  expect(4);
+test('when passed same function twice with same arguments and same target', function() {
+  expect(7);
 
   var bb = new Backburner(['one']);
   var i=0;
   var deferMethod = function(a, b){
     i++;
-    equal(a, 3, 'First argument of only second call is set');
-    equal(b, 2, 'Second argument remains same');
+    equal(a, 1, 'First argument is set twice');
+    equal(b, 2, 'Second argument is set twice');
     equal(this['first'], 1, 'the target property was set');
   };
   var argObj = {'first': 1};
 
   bb.run(function() {
-    bb.deferOnce('one', argObj, deferMethod, 1, 2);
-    bb.deferOnce('one', argObj, deferMethod, 3, 2);
+    bb.defer('one', argObj, deferMethod, 1, 2);
+    bb.defer('one', argObj, deferMethod, 1, 2);
   });
 
-  equal(i, 1, 'function was called once');
+  equal(i, 2, 'function was called twice');
+});
+
+test('when passed same function twice with same target and different arguments', function() {
+  expect(7);
+
+  var bb = new Backburner(['one']),
+      i=0,
+      deferMethod = function(a, b){
+        i++;
+        if(i === 1){
+          equal(a, 1, 'First argument set during first call');
+        } else {
+          equal(a, 3, 'First argument set during second call');
+        }
+        equal(b, 2, 'Second argument remains same');
+        equal(this['first'], 1, 'the target property was set');
+      },
+      argObj = {'first': 1};
+
+  bb.run(function() {
+    bb.defer('one', argObj, deferMethod, 1, 2);
+    bb.defer('one', argObj, deferMethod, 3, 2);
+  });
+
+  equal(i, 2, 'function was called twice');
 });
 
 test('when passed same function twice with different target and different arguments', function() {
@@ -240,33 +224,8 @@ test('when passed same function twice with different target and different argume
   var argObj = {'first': 1};
 
   bb.run(function() {
-    bb.deferOnce('one', {'first': 1}, deferMethod, 1, 2);
-    bb.deferOnce('one', {'first': 1}, deferMethod, 3, 2);
-  });
-
-  equal(i, 2, 'function was called twice');
-});
-
-test('when passed same function with same target after already triggering in current loop (GUID_KEY)', function() {
-  expect(5);
-
-  var bb = new Backburner(['one', 'two'], { 'GUID_KEY': 'GUID_KEY' });
-  var i=0;
-  var deferMethod = function(a){
-    i++;
-    equal(a, i, 'Correct argument is set');
-    equal(this['first'], 1, 'the target property was set');
-  };
-
-  var scheduleMethod = function() {
-    bb.deferOnce('one', argObj, deferMethod, 2);
-  };
-
-  var argObj = {'first': 1, GUID_KEY: '1'};
-
-  bb.run(function() {
-    bb.deferOnce('one', argObj, deferMethod, 1);
-    bb.deferOnce('two', argObj, scheduleMethod);
+    bb.defer('one', {'first': 1}, deferMethod, 1, 2);
+    bb.defer('one', {'first': 1}, deferMethod, 3, 2);
   });
 
   equal(i, 2, 'function was called twice');
@@ -279,10 +238,12 @@ test('onError', function() {
     equal('test error', error.message);
   }
 
-  var bb = new Backburner(['errors'], { onError: onError });
+  var bb = new Backburner(['errors'], {
+    onError: onError
+  });
 
   bb.run(function() {
-    bb.deferOnce('errors', function() {
+    bb.defer('errors', function() {
       throw new Error('test error');
     });
   });
