+++ tests/index.html
@@ -11,7 +11,7 @@
   <script src="qunit.js"></script>
   <script src="/testem.js"></script>
   <script src="loader.js"></script>
-  <script src="backburner.js"></script>
+  <script src="../named-amd/backburner.js"></script>
   <script src="tests.js"></script>
   <script>
     require('backburner-tests');
